fix: 데이터를 index.js에서 다루도록 리팩토링
